moved securityServicePolicyData to the shield stack. add comments to the extended config.
